+++ src/app/admin/admin.routes.ts
@@ -1,10 +1,10 @@
 import { inject } from '@angular/core';
 import { Route } from '@angular/router';
 
-import { AuthService } from '@services/auth.service';
-import { pathNameResolver } from '@resolvers/path-title-resolver.service';
-import { sourceNameResolver } from '@resolvers/source-title-resolver.service';
-import { userNameResolver } from '@resolvers/user-title-resolver.service';
+import { AuthService } from '@services/auth/auth.service';
+import { pathNameResolver } from '@services/path/path-title-resolver.service';
+import { sourceNameResolver } from '@services/source/source-title-resolver.service';
+import { userNameResolver } from '@services/user/user-title-resolver.service';
 
 export default [
   {
@@ -14,32 +14,32 @@ export default [
       {
         path: 'sources',
         title: 'Sources',
-        loadComponent: () => import('./source-list.component'),
+        loadComponent: () => import('./source/source-list.component'),
       },
       {
         path: 'sources/:id',
         title: sourceNameResolver,
-        loadComponent: () => import('./source-edit.component'),
+        loadComponent: () => import('./source/source-edit.component'),
       },
       {
         path: 'paths',
         title: 'Paths',
-        loadComponent: () => import('./path-list.component'),
+        loadComponent: () => import('./path/path-list.component'),
       },
       {
         path: 'paths/:id',
         title: pathNameResolver,
-        loadComponent: () => import('./path-edit.component'),
+        loadComponent: () => import('./path/path-edit.component'),
       },
       {
         path: 'users',
         title: 'Users',
-        loadComponent: () => import('./user-list.component'),
+        loadComponent: () => import('./user/user-list.component'),
       },
       {
         path: 'users/:id',
         title: userNameResolver,
-        loadComponent: () => import('./user-edit.component'),
+        loadComponent: () => import('./user/user-edit.component'),
       },
     ],
     canActivate: [() => inject(AuthService).isLoggedInAsAdmin()],
