+++ src/app/app.component.ts
@@ -1,19 +1,19 @@
 import { Component, OnInit, inject } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 
-import { AuthService } from '@services/auth.service';
+import { AuthService } from '@services/auth/auth.service';
 import { MenuComponent } from './menu/menu.component';
 
 @Component({
-    selector: 'app-root',
-    imports: [MenuComponent, RouterOutlet],
-    template: `
+  selector: 'app-root',
+  imports: [MenuComponent, RouterOutlet],
+  template: `
     <app-menu></app-menu>
     <main>
       <router-outlet></router-outlet>
     </main>
   `,
-    styles: []
+  styles: [],
 })
 export class AppComponent implements OnInit {
   readonly #authService = inject(AuthService);
