+++ src/app/courses/course.routes.ts
@@ -1,8 +1,8 @@
 import { inject } from '@angular/core';
 import { Route } from '@angular/router';
 
-import { AuthService } from '@services/auth.service';
-import { courseTitleResolver } from '@resolvers/course-title-resolver.service';
+import { AuthService } from '@services/auth/auth.service';
+import { courseTitleResolver } from '@services/course/course-title-resolver.service';
 
 export default [
   {
