+++ src/app/menu/menu.component.ts
@@ -3,13 +3,13 @@ import { Router, RouterLink } from '@angular/router';
 
 import { NgbModal, NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
-import { AuthService } from '@services/auth.service';
+import { AuthService } from '@services/auth/auth.service';
 import { LoginComponent } from '@modals/login.component';
 
 @Component({
-    selector: 'app-menu',
-    imports: [NgbModule, RouterLink],
-    template: `
+  selector: 'app-menu',
+  imports: [NgbModule, RouterLink],
+  template: `
     <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
       <a class="navbar-brand" [routerLink]="['/']" id="brand">Training Courses Tracker</a>
 
@@ -40,13 +40,13 @@ import { LoginComponent } from '@modals/login.component';
       </div>
     </nav>
   `,
-    styles: [
-        `
+  styles: [
+    `
       div .nav-item {
         cursor: pointer;
       }
     `,
-    ]
+  ],
 })
 export class MenuComponent {
   readonly #auth = inject(AuthService);
