+++ src/app/shared/services/course/course-title-resolver.service.ts
@@ -3,7 +3,7 @@ import { ActivatedRouteSnapshot, ResolveFn } from '@angular/router';
 
 import { map } from 'rxjs';
 
-import { CoursesService } from '@services/courses.service';
+import { CoursesService } from '@services/course/courses.service';
 
 export const courseTitleResolver: ResolveFn<string> = (route: ActivatedRouteSnapshot) => {
   const id = route.paramMap.get('id');
