+++ src/app/store/index.ts
@@ -1,7 +1,7 @@
 import * as courseState from '@store/course/course.state';
-import * as pathsState from '@store/paths/paths.state';
-import * as sourcesState from '@store/sources/sources.state';
-import * as usersState from '@store/users/users.state';
+import * as pathsState from '@store/path/paths.state';
+import * as sourcesState from '@store/source/sources.state';
+import * as usersState from '@store/user/users.state';
 
 export interface State {
   courses: courseState.State;
